[gatsby-remark-ample] Move utils into utils dir
