+++ plugins/gatsby-remark-ample/package.json
@@ -9,6 +9,6 @@
     "jest": "^25.5.2"
   },
   "scripts": {
-    "test": "jest ./src"
+    "test": "jest ./utils"
   }
 }
