+++ plugins/gatsby-remark-ample/utils/get-key-type.js
