+++ plugins/gatsby-remark-ample/utils/get-options.js
