+++ plugins/gatsby-remark-ample/utils/get-permalink.js
