+++ plugins/gatsby-remark-ample/utils/get-permalink.test.js
