+++ plugins/gatsby-remark-ample/utils/process-image.js
