+++ plugins/gatsby-remark-ample/utils/process-image.test.js
