+++ plugins/gatsby-remark-ample/utils/process-markdown.js
